Adding debug mode for rollup
By default, for compatibility with older Macs, the plugins are transpiled into ES5 Javascript before they are copied to the Plugins folder. This works great, but if you want to try to debug in the Javascript debugger, the transpiled code won't match your code. So for Javascript debugging purposes, use this command instead:

- `npm run autowatch dwertheimer.TaskAutomations -- -debug`